[base] fix for abstract partition manager and its subclasses
Some student reported issues in calling the `Slice` function of abstract partition manager in kv_client_table.hpp because of the use of const. So I modified the abstract_paritition_manager.hpp file so as to avoid you encountering the same problem in the use of const.

Problem:
* no matching function call to `Slice(...)` when using `partition_manager_->Slice(...)` in the KVClientTable class
* this is due to the declaration partition_manager_ in KVClientTable: it is declared as `const AbstractPartitionManager*`, which means only functions with the `const` modifier can be invoked on the class, while I did not give a const modifier to the `Slice` function

ps. Some one discussed with me about the use of const for the `AbstractPartitionManager*` in `KVClientTable`,
he stated that the second const should be removed as we could not expect the pointer stays valid through the life time of a KVClientTable object and suggested the use of shared pointer.
I think this is an interesting argument, so I share with you and leave this for you to judge :)
I would like to reiterate that the codes I give are solely for your reference and you can feel free to make any change if needed.